Refactor NoEncryption scheme.

+++ test/BPlusTree/DeletionTests.cs
@@ -9,7 +9,7 @@ namespace Test
 {
 	public partial class BPlusTreeTests
 	{
-		public Tree<string, int, int> DeleteAndValidate(Tree<string, int, int> tree, int element, bool print = false)
+		public Tree<string, long, int> DeleteAndValidate(Tree<string, long, int> tree, int element, bool print = false)
 		{
 			if (print)
 			{
@@ -33,7 +33,7 @@ namespace Test
 		public void DeleteNotExistingElementTest()
 		{
 			var tree = ConstructTree(
-				new Options<int, int>(
+				new Options<int, long>(
 					new NoEncryptionScheme(),
 					3
 				),
@@ -46,8 +46,8 @@ namespace Test
 		[Fact]
 		public void DeleteEmptyTreeTest()
 		{
-			var tree = new Tree<string, int, int>(
-				new Options<int, int>(
+			var tree = new Tree<string, long, int>(
+				new Options<int, long>(
 					new NoEncryptionScheme(),
 					3
 				)
@@ -60,7 +60,7 @@ namespace Test
 		public void DeleteSingleElementTest()
 		{
 			var tree = ConstructTree(
-				new Options<int, int>(
+				new Options<int, long>(
 					new NoEncryptionScheme(),
 					3
 				),
@@ -74,7 +74,7 @@ namespace Test
 		public void DeleteMultilevelTest()
 		{
 			var tree = ConstructTree(
-				new Options<int, int>(
+				new Options<int, long>(
 					new NoEncryptionScheme(),
 					3
 				),
@@ -91,7 +91,7 @@ namespace Test
 		public void DeleteThenInsertTest()
 		{
 			var tree = ConstructTree(
-				new Options<int, int>(
+				new Options<int, long>(
 					new NoEncryptionScheme(),
 					3
 				),
@@ -112,7 +112,7 @@ namespace Test
 		public void DeleteLargestElement()
 		{
 			var tree = ConstructTree(
-				new Options<int, int>(
+				new Options<int, long>(
 					new NoEncryptionScheme(),
 					3
 				),
@@ -126,7 +126,7 @@ namespace Test
 		public void DeleteWithMergeToRight()
 		{
 			var tree = ConstructTree(
-				new Options<int, int>(
+				new Options<int, long>(
 					new NoEncryptionScheme(),
 					3
 				),
@@ -141,7 +141,7 @@ namespace Test
 		public void DeleteWithBorrowFromLeft()
 		{
 			var tree = ConstructTree(
-				new Options<int, int>(
+				new Options<int, long>(
 					new NoEncryptionScheme(),
 					3
 				),
@@ -157,7 +157,7 @@ namespace Test
 		public void DeleteWithBorrowFromRight()
 		{
 			var tree = ConstructTree(
-				new Options<int, int>(
+				new Options<int, long>(
 					new NoEncryptionScheme(),
 					3
 				),
@@ -183,7 +183,7 @@ namespace Test
 						.Distinct()
 						.ToList();
 				var tree = ConstructTree(
-					new Options<int, int>(
+					new Options<int, long>(
 						new NoEncryptionScheme(),
 						i
 					),
